requirements.txt
opencv-python~=4.10.0.84
numpy~=2.2.1
scipy~=1.14.1
pillow~=11.0.0